refactor(browse): own a Results component per content type
Replace the `useResults` hook stored on each `ContentType` with a
`Results` component, and render `<activeType.Results />`. The three
implementations call different numbers of query hooks (all=3, items=2,
spells=1), so the old single `<Results>` instance relied on
`key={activeType.id}` to remount on type switch and avoid a
rules-of-hooks violation. That key was load-bearing for correctness yet
read like a routine list key — a regression footgun.

Rendering a different component type per content type makes React remount
on switch via normal reconciliation, so the invariant is structural and
the key is gone. Each type owns its data fetching and delegates rendering
to a shared `ResultsView`; the `type.id === "all"` row-meta special-case
collapses to a `row.kindLabel` check (equivalent — only `all` sets it).

Relocate the whole content-type module from `src/api/content-types/` to
`src/views/browse/`: it carries presentation (ResultsView + CSS) and
imports `lib/ui`, so it belongs in the view layer alongside its only
consumer, mirroring the existing `src/views/reference/` feature folder.
`src/api/` returns to data-only.

Add tests for the All->Spells->All round-trip (guards the hook-count
invariant) and the per-type empty-state copy.
