+++ src/views/BrowseApiModal.module.css
@@ -55,79 +55,6 @@
   display: block;
 }
 
-.results {
-  flex: 1;
-  min-height: 0;
-  display: flex;
-  flex-direction: column;
-  overflow-y: auto;
-  padding: var(--space-1) 0;
-}
-
-.row {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-  padding: var(--space-2) var(--space-4);
-  border: 0;
-  background: transparent;
-  text-align: left;
-  font: inherit;
-  font-family: var(--font-body);
-  cursor: pointer;
-  border-bottom: 1px solid var(--color-border);
-}
-
-.row:last-child {
-  border-bottom: 0;
-}
-
-.row:hover:not(:disabled) {
-  background: var(--color-surface-2);
-}
-
-.row:focus-visible {
-  outline: 2px solid var(--color-focus-ring);
-  outline-offset: -2px;
-}
-
-.row:disabled {
-  cursor: not-allowed;
-  opacity: 0.5;
-}
-
-.rowName {
-  font-weight: 600;
-}
-
-.rowMeta {
-  font-size: var(--fs-sm);
-  color: var(--color-text-muted);
-}
-
-.state {
-  padding: var(--space-5);
-  text-align: center;
-  color: var(--color-text-muted);
-}
-
-.stateError {
-  color: var(--color-danger);
-  background: var(--color-danger-bg-hover);
-}
-
-.statePane {
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.errorActions {
-  margin-top: var(--space-2);
-}
-
 .footer {
   margin: 0;
   padding: var(--space-3) var(--space-4);
